read down //
1 balance
2 map
3 stats
problem that hp and damage are not float

+++ Assets/Scripts/UnitStats.cs
@@ -6,8 +6,8 @@ using UnityEngine.UI;
 
 public class UnitStats : MonoBehaviour
 {
-    public int maxHP;
-    public int damage;
+    public float maxHP;
+    public float damage;
     public float attackSpeed; 
     public float speed;
     public float lifeTime;
